DX12: small indices support

+++ DX12/D3D12Research/D3D12/Graphics/DynamicResourceAllocator.cpp
@@ -8,9 +8,9 @@ DynamicResourceAllocator::DynamicResourceAllocator(DynamicAllocationManager* pPa
 {
 }
 
-DynamicAllocation DynamicResourceAllocator::Allocate(uint64_t size, int alignment)
+DynamicAllocation DynamicResourceAllocator::Allocate(size_t size, int alignment)
 {
-	int bufferSize = (size + (alignment - 1)) & ~(alignment - 1);
+	size_t bufferSize = (size + ((int64_t)alignment - 1)) & ~(alignment - 1);
 	DynamicAllocation allocation;
 	allocation.Size = bufferSize;
 
@@ -67,7 +67,7 @@ DynamicAllocationManager::~DynamicAllocationManager()
 {
 }
 
-AllocationPage* DynamicAllocationManager::AllocatePage(uint64_t size)
+AllocationPage* DynamicAllocationManager::AllocatePage(size_t size)
 {
 	std::scoped_lock lockGuard(m_PageMutex);
 
@@ -86,7 +86,7 @@ AllocationPage* DynamicAllocationManager::AllocatePage(uint64_t size)
 	return pPage;
 }
 
-AllocationPage* DynamicAllocationManager::CreateNewPage(uint64_t size)
+AllocationPage* DynamicAllocationManager::CreateNewPage(size_t size)
 {
 	AllocationPage* pNewPage = new AllocationPage();
 	pNewPage->Create(m_pGraphics, size, true);
