+++ DX12/D3D12Research/D3D12/Graphics/DynamicResourceAllocator.h
@@ -7,8 +7,8 @@ struct DynamicAllocation
 {
 	GraphicsBuffer* pBackingResource{ nullptr };
 	D3D12_GPU_VIRTUAL_ADDRESS GpuHandle{ 0 };
-	int Offset{ 0 };
-	int Size{ 0 };
+	size_t Offset{ 0 };
+	size_t Size{ 0 };
 	void* pMappedMemory{ nullptr };
 };
 
@@ -20,8 +20,8 @@ public:
 	DynamicAllocationManager(Graphics* pGraphics);
 	~DynamicAllocationManager();
 
-	AllocationPage* AllocatePage(uint64_t size);
-	AllocationPage* CreateNewPage(uint64_t size);
+	AllocationPage* AllocatePage(size_t size);
+	AllocationPage* CreateNewPage(size_t size);
 
 	void FreePages(uint64_t fenceValue, const std::vector<AllocationPage*> pPages);
 	void FreeLargePages(uint64_t fenceValue, const std::vector<AllocationPage*> pLargePages);
@@ -39,16 +39,16 @@ class DynamicResourceAllocator
 public:
 	DynamicResourceAllocator(DynamicAllocationManager* pPageManager);
 	
-	DynamicAllocation Allocate(uint64_t size, int alignment = 256);
+	DynamicAllocation Allocate(size_t size, int alignment = 256);
 	void Free(uint64_t fenceValue);
 
 private:
-	constexpr static uint64_t PAGE_SIZE = 0xFFFF;
+	constexpr static size_t PAGE_SIZE = 0xFFFF;
 
 	DynamicAllocationManager* m_pPageManager;
 
 	AllocationPage* m_pCurrentPage{ nullptr };
-	int m_CurrentOffset{ 0 };
+	size_t m_CurrentOffset{ 0 };
 	std::vector<AllocationPage*> m_UsedPages;
 	std::vector<AllocationPage*> m_UsedLargePages;
 };
